+++ packages/polaris-viz/src/components/SimpleBarChart/stories/WithTrendIndicators.stories.tsx
@@ -54,6 +54,10 @@ WithTrendIndicators.args = {
             direction: 'upward',
             trend: 'positive',
           },
+          3: {
+            direction: 'upward',
+            trend: 'positive',
+          },
         },
       },
     },