+++ packages/polaris-viz-core/src/constants.ts
@@ -504,7 +504,6 @@ export const ELLIPSIS = '…';
 export const HORIZONTAL_LABEL_MIN_WIDTH = 46;
 export const HORIZONTAL_LABEL_TARGET_HEIGHT = 80;
 export const DIAGONAL_LABEL_MIN_WIDTH = 30;
-export const DONUT_CHART_MAX_SERIES_COUNT = 5;
 export const MAX_DIAGONAL_LABEL_WIDTH = 100;
 // Visible height of a 100px wide label at 45deg
 export const MAX_DIAGONAL_VISIBLE_HEIGHT = 80;