+++ packages/polaris-viz/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- Changed `<LineChart />` and `vertical` `<BarChart />` to use a react portal to allow tooltips to render outside the bounds of the chart.
 
 ## [9.9.0] - 2023-07-26